refactor: split shared session ops from session_glue.rs
Move 14 shared-session replication functions into shared_ops.rs
(477 lines). session_glue.rs drops from 3561 to 3085 lines.

Moved: publish/remove/demote shared sessions, cross-scope lookups,
synced entry construction, reverse prewarm/install.

Closes #369
